+++ src/App.js
@@ -1,11 +1,12 @@
-import { useState } from 'react';
-import Banner from './components/Banner/Banner.jsx';
-import { NavigationContext } from './store/navigation-context.jsx';
+import { useState } from "react";
+import Banner from "./components/Banner/Banner.jsx";
+import Main from './components/Main/Main.jsx';
+import { NavigationContext } from "./store/navigation-context.jsx";
 
 import "./App.css";
 
 function App() {
-  const [selectedNavItem, setSelectedNavItem] = useState('Metrics');
+  const [selectedNavItem, setSelectedNavItem] = useState("Metrics");
 
   function selectNavItemHandler(selectedItem) {
     setSelectedNavItem(() => selectedItem);
@@ -15,13 +16,16 @@ function App() {
 
   const ctxNavigation = {
     selectedNavItem: selectedNavItem,
-    selectNavItemHandler: selectNavItemHandler
+    selectNavItemHandler: selectNavItemHandler,
   };
 
   return (
-    <NavigationContext.Provider value={ctxNavigation}>
-      <Banner />
-    </NavigationContext.Provider>
+    <>
+      <NavigationContext.Provider value={ctxNavigation}>
+        <Banner />
+        <Main />
+      </NavigationContext.Provider>      
+    </>
   );
 }