+++ src/components/Banner/Logo/Logo.styles.jsx
@@ -3,6 +3,7 @@ import { styled } from "styled-components";
 export const $HeaderLogo = styled.img`
   height: 90%;
   margin: 5px 100px;
+  user-select: none;
 `;
 
 export default $HeaderLogo;