+++ src/components/Main/Main.jsx
@@ -1,0 +1,10 @@
+import Metrics from './Metrics/Metrics.jsx';
+import $Main from './Main.styles.jsx';
+
+export default function Main() {
+    return (
+        <$Main>
+            <Metrics />
+        </$Main>
+    );
+}