+++ src/components/Main/Metrics/MetricsStepper/MetricsStepper.styles.jsx
@@ -1,0 +1,11 @@
+import { styled } from "styled-components";
+
+export const $MetricsStepper = styled.div`
+  margin: 20px 100px 30px 0;
+  display: flex;
+  justify-content: space-evenly;
+`;
+
+
+
+