+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.13 (2024-05-10)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/3)
+* Create metrics stepper
+* Add application colors to separate jsx file
+* Create and style company items
+* Create metrics-context and implement it to proper components
+
 0.12 (2024-05-04)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/2)